Fix test failure from r351495
The test has problems due to some platforms having a different type for
ptrdiff_t, so the error message is different.  The error message doesn't
matter to the test for anything other than an incompatible intger to
pointer conversion, so this patch removes the integral type from the
expected message.

Change-Id: I80e786f9b80268163813774bbf25a9ca25b6c60c

git-svn-id: http://llvm.org/svn/llvm-project/cfe/trunk@351550 91177308-0d34-0410-b5e6-96231b3b80d8